Bug Fixes and Redesigns
Fixed bugs with the touch and the selection fo the cells
Passed a weak reference of the VC to the cell so we can control if the status bar can be shown or hidden while animated.
Also the control of isOpen on the CollectionVC is now assignable for the cell.  We will have to think of a better way to design this in the future with a better refactor.